Move to CppUTestExt & put into Dos build
